Fix strange PSS tracking issue.
Turns out that srslte_pss_synch_t struct keeps a moving avg of the PSS
convolution and just as it had found a solid peak location, I was
"recentering" the input, which then consistently caused it to return a
very low average for the new peak location.

The solution was to nuke that convolution average cache when I start tracking.