Modify how PSS decides when to forward frames
Previously, frames were _not_ forwarded while building up the tracking
score (before tracking), and _were_ forwarded while decrementing the tracking score. The
point was less thrashing of various caches, but in practice, this led to
a lot of garbage making its way to the SSS block when input was of poor
quality. Oh, and crashes. Now, tracking score and state within PSS is
used only skipping the PSR threshold test when signal quality is good.
decr_score is now reset_score, because it is safer to drop out of
tracking. This may not have been necessary and may hurt performance, so
should reevaluate resetting vs decrementing the tracking score at a
later date.